+++ CopilotKit/packages/react-ui/src/components/chat/Markdown.tsx
@@ -28,6 +28,18 @@ const components: Components = {
   p({ children }) {
     return <p>{children}</p>;
   },
+  a({ children, ...props }) {
+    return (
+      <a
+        style={{ color: "blue", textDecoration: "underline" }}
+        {...props}
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        {children}
+      </a>
+    );
+  },
   code({ children, className, inline, ...props }) {
     if (children.length) {
       if (children[0] == "▍") {